refactor: split report query family
